pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "humanoidverse"
version = "0.0.1"
description = "New Version of Humanoidverse"
authors = [
    { name = "Qiwei Wu", email = "nathan.wuqw@gmail.com" },
]
requires-python = ">=3.8"
# These are the core dependencies that are common to both environments
dependencies = [
    "hydra-core>=1.2.0",
    "numpy",
    "rich",
    "ipdb",
    "matplotlib",
    "termcolor",
    "wandb",
    "plotly",
    "tqdm",
    "loguru",
    "meshcat",
    "pynput",
    "scipy",
    "tensorboard",
    "onnx",
    "onnxruntime",
    "opencv-python",
    "easydict",
    "lxml"
]
